refactor: unify middleware application for API routes and tests
